+++ docs_src/docs/CHANGELOG.md
@@ -7,12 +7,17 @@
 - [`monad_std.iter.IterMeta`][monad_std.iter.iter.IterMeta]:
     - [`map_while`][monad_std.iter.iter.IterMeta.map_while]: Yields elements based on both a predicate and maps.
     - [`map_windows`][monad_std.iter.iter.IterMeta.map_windows]: A map but with fixed-sized windows.
-    - [`max`][monad_std.iter.iter.IterMeta.max], [`min`][monad_std.iter.iter.IterMeta.min]:
+    - [`max`][monad_std.iter.iter.IterMeta.max], [`max_by`][monad_std.iter.iter.IterMeta.max_by],
+      [`max_by_key`][monad_std.iter.iter.IterMeta.max_by_key], [`min`][monad_std.iter.iter.IterMeta.min]
+      [`min_by`][monad_std.iter.iter.IterMeta.min_by], [`min_by_key`][monad_std.iter.iter.IterMeta.min_by_key]:
       Get the maximum / minimum value.
     - [`collect_to_xxx`][monad_std.iter.iter.IterMeta.collect_to_seq]:
       Collect the iterator into a mutable and extendable item.<br />
       Note that in Python it's not possible to actually *change* a `str`, so `collect_to_string` is not provided.
     - [`collect_set`][monad_std.iter.iter.IterMeta.collect_set]: Collect the iterator into a hashset.
+- [`monad_std.typedef`](./api_document/typedef.md): Internal type definition.
+- [`monad_std.utils`](./api_document/utils/index.md):
+  Utility used inside and outside (recommend, but not forced) the library.
 
 **Impl Change**